[core] library: export: move notification code to a dedicated file
