+++ projects/core/library/export.py
@@ -16,21 +16,16 @@ from typing import Any
 import click
 import yaml
 
-from projects.caliper.orchestration.export import run_from_orchestration_config
-from projects.core.ci_entrypoint.prepare_ci import CI_METADATA_DIRNAME
+from projects.caliper.orchestration.export import (
+    run_from_orchestration_config,
+)
 from projects.core.library import ci as ci_lib
 from projects.core.library import config, env, run
-
-
-class StepStatus(StrEnum):
-    """Status of a step execution."""
-
-    SUCCESS = "success"
-    FAILURE = "failure"
-    ONGOING = "ongoing"
-    UNKNOWN = "unknown"
-    WARNING = "warning"
-
+from projects.core.library.export_notifications import (
+    _check_job_shutdown_status,
+    _create_mlflow_file_url_for_step,
+    send_notification,
+)
 
 logger = logging.getLogger(__name__)
 
@@ -96,345 +91,6 @@ def _update_fjob_export_status(status: dict):
         # Restore KUBECONFIG if it was set
         if original_kubeconfig is not None:
             os.environ["KUBECONFIG"] = original_kubeconfig
-
-
-def send_notification(
-    artifact_dir: Path | None,
-    status: dict[str, Any],
-    notification_provider=None,
-    dry_run: bool = False,
-) -> bool:
-    """Send job completion notifications based on caliper export status.
-
-    Args:
-        artifact_dir: Directory to browse to find the artifacts
-        status: Caliper export status object containing backend results and metadata
-        notification_provider: Optional per-project SlackNotificationProvider instance
-        dry_run: If True, only build and log notification content without sending
-
-    Returns:
-        bool: True if notifications were sent successfully, False otherwise
-    """
-    # Extract notification parameters from status object
-
-    project = config.project.get_config("project.name")
-    finish_reason = _extract_finish_reason_from_status(status)
-    duration_str = _extract_duration_from_status(status)
-
-    # Build enhanced notification with fournos job info and artifact links
-    notification_status, notification_success = _build_enhanced_notification(
-        artifact_dir, project, finish_reason, duration_str, status
-    )
-
-    # Send actual notifications
-    if dry_run:
-        logger.info(f"DRY RUN: Would send notification: {project} {finish_reason}{duration_str}")
-        logger.info(f"DRY RUN: Notification content:\n{notification_status}")
-        return notification_success
-    else:
-        logger.info(f"Sending notification: {project} {finish_reason}{duration_str}")
-
-    # Write notification to file for GitHub pickup (always generate, even in dry-run)
-    try:
-        if env.ARTIFACT_DIR:
-            notification_file = Path(env.ARTIFACT_DIR) / "NOTIFICATION-github.md"
-            with open(notification_file, "w", encoding="utf-8") as f:
-                f.write(notification_status + "\n")
-            if dry_run:
-                logger.info(f"DRY RUN: Generated notification file {notification_file}")
-            else:
-                logger.info(f"Wrote export notification file {notification_file}")
-        else:
-            logger.warning("ARTIFACT_DIR not available, skipping notification file")
-    except Exception as e:
-        logger.exception(f"Failed to write notification file: {e}")
-
-    # Actually send notification through GitHub API
-    try:
-        from projects.core.notifications.send import send_notification as send_github_notification
-
-        # Get notification vault from configuration
-        notification_vault = None
-        try:
-            notification_config = config.project.get_config("caliper.export.notifications", {})
-            notification_vault = notification_config.get("vault")
-            if notification_vault:
-                logger.info(f"Using notification vault from config: {notification_vault}")
-        except Exception as e:
-            logger.warning(f"Failed to get notification vault from config: {e}")
-
-        success = send_github_notification(
-            message=notification_status,
-            github=True,
-            slack=False,
-            dry_run=dry_run,
-            notification_vault=notification_vault,
-        )
-        if success:
-            logger.info("Successfully sent GitHub notification")
-        else:
-            logger.error("GitHub notification sending failed")
-            notification_success = False
-    except Exception as e:
-        logger.error(f"Failed to send GitHub notification: {e}")
-        notification_success = False
-
-    # Per-project Slack notification via provider
-    if notification_provider:
-        if not dry_run:
-            try:
-                from projects.core.notifications.provider import NotificationContext
-
-                artifact_dir = Path(env.ARTIFACT_DIR) if env.ARTIFACT_DIR else None
-                context = NotificationContext(
-                    status=status,
-                    finish_reason=str(finish_reason),
-                    project_name=project or "unknown",
-                    pr_number=os.environ.get("PULL_NUMBER"),
-                    job_type=os.environ.get("JOB_TYPE"),
-                    artifact_dir=artifact_dir,
-                )
-                ok = notification_provider.notify(context)
-                if ok:
-                    logger.info("Successfully sent per-project Slack notification")
-                else:
-                    logger.warning("Per-project Slack notification failed")
-                    notification_success = False
-            except Exception as e:
-                logger.warning(f"Failed to send per-project Slack notification: {e}")
-                notification_success = False
-        else:
-            logger.info("DRY RUN: Would send per-project Slack notification")
-
-    return notification_success
-
-
-def _get_project_and_args(project: str) -> tuple[str, str]:
-    """Extract project name and args from fournos job or config."""
-    fjob_project = project
-    fjob_args_str = ""
-
-    try:
-        metadata_dir = ci_lib.get_ci_metadata_dir()
-        fournos_fjob_path = metadata_dir / "fournos_fjob.yaml"
-        if not fournos_fjob_path.exists():
-            return fjob_project, fjob_args_str
-
-        with open(fournos_fjob_path, encoding="utf-8") as f:
-            fjob_data = yaml.safe_load(f)
-
-        display_name = fjob_data.get("spec", {}).get("displayName", "")
-        if not display_name:
-            return fjob_project, fjob_args_str
-
-        parts = display_name.split()
-        if not parts:
-            return fjob_project, fjob_args_str
-
-        fjob_project = parts[0]
-        fjob_args_str = " ".join(parts[1:]) if len(parts) > 1 else ""
-    except Exception as e:
-        logger.warning(f"Failed to read fournos job for project/args: {e}")
-
-    if fjob_args_str:
-        return fjob_project, fjob_args_str
-
-    try:
-        from projects.core.library import config
-
-        job_args = config.project.get_config("ci_job.args", [], warn=False)
-        fjob_args_str = " ".join(job_args) if job_args else ""
-    except Exception as e:
-        logger.warning(f"Failed to get args from config: {e}")
-
-    return fjob_project, fjob_args_str
-
-
-def _get_execution_engine_config() -> str | None:
-    """Read and format execution engine configuration."""
-    try:
-        metadata_dir = ci_lib.get_ci_metadata_dir()
-        fournos_fjob_path = metadata_dir / "fournos_fjob.yaml"
-        if not fournos_fjob_path.exists():
-            return None
-
-        with open(fournos_fjob_path, encoding="utf-8") as f:
-            fjob_data = yaml.safe_load(f)
-
-        execution_engine = fjob_data.get("spec", {}).get("executionEngine", {})
-        if not execution_engine:
-            return None
-
-        engine_yaml = yaml.dump(execution_engine, default_flow_style=False, sort_keys=True)
-        return f"```yaml\n{engine_yaml.strip()}\n```"
-    except Exception as e:
-        logger.warning(f"Failed to read fournos job config: {e}")
-        return None
-
-
-def _check_job_shutdown_status() -> dict[str, Any] | None:
-    """Check if the job has been aborted via spec.shutdown field."""
-    try:
-        metadata_dir = ci_lib.get_ci_metadata_dir()
-        fournos_fjob_path = metadata_dir / "fournos_fjob.yaml"
-        if not fournos_fjob_path.exists():
-            return None
-
-        with open(fournos_fjob_path, encoding="utf-8") as f:
-            fjob_data = yaml.safe_load(f)
-
-        shutdown_value = fjob_data.get("spec", {}).get("shutdown")
-        if shutdown_value:
-            return {
-                "shutdown_detected": True,
-                "shutdown_value": shutdown_value,
-                "is_aborted": shutdown_value.lower() == "stop",
-            }
-
-        return {"shutdown_detected": False, "shutdown_value": None, "is_aborted": False}
-    except Exception as e:
-        logger.warning(f"Failed to check job shutdown status: {e}")
-        return None
-
-
-def _extract_artifact_links(status: dict[str, Any]) -> tuple[list[str], str | None]:
-    """Extract artifact links and MLflow URL from status."""
-    artifact_links = []
-    mlflow_run_url = None
-
-    caliper_export = status.get("caliper_artifacts_export", {})
-    backends = caliper_export.get("backends", {})
-
-    for backend_name, backend_result in backends.items():
-        if not isinstance(backend_result, dict):
-            continue
-
-        if backend_result.get("experiment_url"):
-            artifact_links.append(
-                f"[{backend_name} Experiment]({backend_result['experiment_url']})"
-            )
-
-        if backend_result.get("run_url"):
-            mlflow_run_url = backend_result["run_url"]
-            artifact_links.append(f"[{backend_name} Results]({mlflow_run_url})")
-        elif backend_result.get("artifact_url"):
-            artifact_links.append(f"[{backend_name} Artifacts]({backend_result['artifact_url']})")
-        elif backend_result.get("dashboard_url"):
-            artifact_links.append(f"[{backend_name} Dashboard]({backend_result['dashboard_url']})")
-
-    if status.get("artifact_url"):
-        artifact_links.append(f"[Artifacts]({status['artifact_url']})")
-
-    return artifact_links, mlflow_run_url
-
-
-def _create_mlflow_url(mlflow_run_url: str, step_dir_name: str) -> str | None:
-    """Create MLflow URL for step logs."""
-
-    if not mlflow_run_url:
-        return f"BASE_URL_MISSING/{step_dir_name}"
-
-    if "/artifacts" not in mlflow_run_url:
-        logger.warning(f"Unexpected MLflow URL format: {mlflow_run_url}")
-        return None
-
-    if "#" in mlflow_run_url:
-        base_domain, hash_fragment = mlflow_run_url.split("#", 1)
-        if "/artifacts" not in hash_fragment:
-            raise ValueError("Artifacts not found in hash fragment")
-
-        hash_base, params = hash_fragment.split("/artifacts", 1)
-        workspace_param = params if "?workspace=" in params else ""
-        return f"{base_domain}#{hash_base}/artifacts/{step_dir_name}/run.log{workspace_param}"
-    else:
-        base_url, params = mlflow_run_url.split("/artifacts", 1)
-        workspace_param = params if "?workspace=" in params else ""
-        return f"{base_url}/artifacts/{step_dir_name}/run.log{workspace_param}"
-
-
-def _create_mlflow_step_url(mlflow_run_url: str, step_dir_name: str) -> str | None:
-    """Create MLflow URL for step directory (for file access)."""
-    if "/artifacts" not in mlflow_run_url:
-        logger.warning(f"Unexpected MLflow URL format: {mlflow_run_url}")
-        return None
-
-    if "#" in mlflow_run_url:
-        base_domain, hash_fragment = mlflow_run_url.split("#", 1)
-        if "/artifacts" not in hash_fragment:
-            raise ValueError("Artifacts not found in hash fragment")
-
-        hash_base, artifacts_part = hash_fragment.split("/artifacts", 1)
-        # Extract workspace parameter if present, ignoring existing path
-        workspace_param = ""
-        if "?workspace=" in artifacts_part:
-            workspace_param = artifacts_part[artifacts_part.find("?") :]
-        return f"{base_domain}#{hash_base}/artifacts/{step_dir_name}{workspace_param}"
-    else:
-        base_url, artifacts_part = mlflow_run_url.split("/artifacts", 1)
-        # Extract workspace parameter if present, ignoring existing path
-        workspace_param = ""
-        if "?workspace=" in artifacts_part:
-            workspace_param = artifacts_part[artifacts_part.find("?") :]
-        return f"{base_url}/artifacts/{step_dir_name}{workspace_param}"
-
-
-def _create_mlflow_file_url_for_step(
-    mlflow_run_url: str, step_dir_name: str, file_path: str
-) -> str:
-    """Create MLflow URL for a specific file within a step directory.
-
-    Args:
-        mlflow_run_url: Base MLflow run URL
-        step_dir_name: Name of the step directory
-        file_path: Relative path to file from step directory
-
-    Returns:
-        Full MLflow URL to the file
-
-    Raises:
-        ValueError: If URL format is unexpected
-    """
-    if "/artifacts" not in mlflow_run_url:
-        raise ValueError(f"Unexpected MLflow URL format: {mlflow_run_url}")
-
-    # Clean file path
-    file_clean = file_path.lstrip("/")
-
-    if "#" in mlflow_run_url:
-        base_domain, hash_fragment = mlflow_run_url.split("#", 1)
-        if "/artifacts" not in hash_fragment:
-            raise ValueError("Artifacts not found in hash fragment")
-
-        hash_base, artifacts_part = hash_fragment.split("/artifacts", 1)
-        # Extract workspace parameter if present, ignoring existing path
-        workspace_param = ""
-        if "?workspace=" in artifacts_part:
-            workspace_param = artifacts_part[artifacts_part.find("?") :]
-        return f"{base_domain}#{hash_base}/artifacts/{step_dir_name}/{file_clean}{workspace_param}"
-    else:
-        base_url, artifacts_part = mlflow_run_url.split("/artifacts", 1)
-        # Extract workspace parameter if present, ignoring existing path
-        workspace_param = ""
-        if "?workspace=" in artifacts_part:
-            workspace_param = artifacts_part[artifacts_part.find("?") :]
-        return f"{base_url}/artifacts/{step_dir_name}/{file_clean}{workspace_param}"
-
-
-def _read_step_duration(step_dir: Path) -> str:
-    """Read step duration from timing file."""
-    timing_file = step_dir / CI_METADATA_DIRNAME / "test_duration.yaml"
-    if not timing_file.exists():
-        return ""
-
-    try:
-        with open(timing_file, encoding="utf-8") as f:
-            timing_data = yaml.safe_load(f)
-
-        formatted_duration = timing_data.get("duration", {}).get("formatted")
-        return formatted_duration or ""
-    except Exception as timing_error:
-        logger.warning(f"Failed to read timing file {timing_file}: {timing_error}")
-        return ""
 
 
 def _process_caliper_postprocess_status(
@@ -507,489 +163,6 @@ def _process_caliper_postprocess_status(
         except Exception as e:
             logger.error(f"Failed to process caliper postprocess status file {status_file}: {e}")
             raise
-
-
-def _process_notification_files(step_dir: Path) -> None:
-    """Process notification files from step directory."""
-    notifications_dir = step_dir / CI_METADATA_DIRNAME / "notifications"
-    if not (notifications_dir.exists() and notifications_dir.is_dir()):
-        return []
-
-    import re
-
-    notifications_from_files = []
-    for notification_file in sorted(notifications_dir.glob("*.txt")):
-        with open(notification_file, encoding="utf-8") as f:
-            content = f.read().strip()
-
-        if not content:
-            continue
-
-        subtitle = notification_file.stem.replace("__", " ").replace("_", " ").title()
-        subtitle = re.sub(r"^\d+\s+", "", subtitle)
-        notifications_from_files.append(f"##### {subtitle}")
-
-        for line in content.splitlines():
-            notifications_from_files.append(f"> {line}")
-
-    return notifications_from_files
-
-
-def _process_step_status(artifact_dir: Path, mlflow_run_url: str) -> list[str]:
-    """Process step logs from parent directory."""
-
-    if not mlflow_run_url:
-        logging.warning("mlflow_run_url not set. Will generate dummy links.")
-
-    step_status = []
-
-    current_step_name = Path(env.BASE_ARTIFACT_DIR).name
-
-    for step_dir in sorted(artifact_dir.iterdir()):
-        if not step_dir.is_dir():
-            continue
-        if step_dir.name.startswith("."):
-            continue
-
-        run_log = step_dir / "run.log"
-        if not run_log.exists():
-            continue
-
-        mlflow_log_url = _create_mlflow_url(mlflow_run_url, step_dir.name)
-        if not mlflow_log_url:
-            mlflow_log_url = "NO_URL"
-
-        step_name = step_dir.name.replace("__", " ").replace("_", " ").title()
-        duration_str = _read_step_duration(step_dir)
-        exit_status_emoji, exit_status = _read_step_exit_status(step_dir, current_step_name)
-
-        step_status.append("")
-        if duration_str:
-            step_status.append(
-                f"#### {exit_status_emoji} [{step_name}]({mlflow_log_url}) `{duration_str}`"
-            )
-        else:
-            step_status.append(f"#### {exit_status_emoji} [{step_name}]({mlflow_log_url})")
-
-        step_status.extend(_process_notification_files(step_dir))
-
-        step_details = _process_step_details(step_dir, mlflow_run_url)
-
-        # Add test execution info for this step
-        if step_details["test_labels_info"]:
-            # Add Test Execution Overview if we have any step info
-            step_status.append("")
-            step_status.append("**Test Execution Overview**")
-
-            step_status.extend([f"* {info}" for info in step_details["test_labels_info"]])
-
-        # Add postprocess info for this step
-        if step_details["postprocess_info"]:
-            step_status.extend([f"* {info}" for info in step_details["postprocess_info"]])
-
-    return step_status
-
-
-def _process_postprocess_status(mlflow_run_url: str | None = None) -> list[str]:
-    """Process post-processing status from all step directories."""
-    if not mlflow_run_url:
-        return []
-
-    postprocess_links = []
-    parent_dir = Path(env.BASE_ARTIFACT_DIR).parent
-
-    for step_dir in sorted(parent_dir.iterdir()):
-        if not step_dir.is_dir():
-            continue
-        if step_dir.name.startswith("."):
-            continue
-
-        try:
-            _process_caliper_postprocess_status(step_dir, postprocess_links, mlflow_run_url)
-        except Exception as e:
-            logger.error(f"Failed to process postprocess status for {step_dir.name}: {e}")
-            raise
-
-    return postprocess_links
-
-
-def _read_step_exit_status(
-    step_dir: Path, current_step_name: str | None = None
-) -> tuple[str, StepStatus]:
-    """Read exit status from step directory and return emoji and status enum."""
-    try:
-        exit_status_file = step_dir / CI_METADATA_DIRNAME / "exit_status.yaml"
-        if not exit_status_file.exists():
-            # Check if this is the current ongoing step
-            if current_step_name and step_dir.name == current_step_name:
-                return "🔄", StepStatus.ONGOING  # Ongoing step
-            return "❓", StepStatus.UNKNOWN  # Unknown status if file doesn't exist
-
-        with open(exit_status_file, encoding="utf-8") as f:
-            exit_data = yaml.safe_load(f)
-
-        return_code = exit_data.get("return_code")
-        if return_code is None or return_code == 0:
-            return "✅", StepStatus.SUCCESS
-        else:
-            return "❌", StepStatus.FAILURE
-    except Exception as e:
-        logger.warning(f"Failed to read exit status from {step_dir}: {e}")
-        # Check if this is the current ongoing step even on error
-        if current_step_name and step_dir.name == current_step_name:
-            return "🔄", StepStatus.ONGOING  # Ongoing step
-        return "❓", StepStatus.UNKNOWN  # Unknown status on error
-
-
-def _check_postprocess_warnings(step_dir: Path) -> StepStatus:
-    """Check for warning status in postprocess status file."""
-
-    status = StepStatus.SUCCESS  # No postprocess warning/error, assume no warnings
-    for status_file in step_dir.glob("**/postprocess_status.yaml"):
-        try:
-            with open(status_file, encoding="utf-8") as f:
-                status_data = yaml.safe_load(f)
-        except Exception as e:
-            logging.error(f"Failed to read {status_file} as yaml: {e}")
-            status = StepStatus.WARNING
-            continue
-
-        if not status_data:
-            continue
-
-        # Check top-level success field for warning value
-        success_value = status_data.get("success")
-        if success_value == "warning":
-            logging.warning(
-                f"Post-process warning detected in {status_file}, setting the WARNING flag"
-            )
-            status = StepStatus.WARNING
-
-        if success_value in ("failure", "error"):
-            logging.error(f"Post-process {success_value} detected, raising the FAILURE flag")
-            return StepStatus.FAILURE
-
-    return status
-
-
-def _get_overall_status_from_steps(artifact_dir) -> str:
-    """Check all step exit statuses and return overall status emoji."""
-    try:
-        current_step_name = Path(env.BASE_ARTIFACT_DIR).name
-
-        step_statuses = []
-
-        for step_dir in sorted(artifact_dir.iterdir()):
-            if not step_dir.is_dir():
-                continue
-            if step_dir.name.startswith("."):
-                continue
-
-            # Only check directories that have run.log (actual steps)
-            run_log = step_dir / "run.log"
-            if not run_log.exists():
-                continue
-
-            _emoji, status = _read_step_exit_status(step_dir, current_step_name)
-
-            step_statuses.append(status)
-
-            # Check for postprocess warnings in this step (always check, regardless of exit status)
-            postprocess_status = _check_postprocess_warnings(step_dir)
-            step_statuses.append(postprocess_status)
-
-        # Priority: failure > ongoing > warning > unknown > success
-        if StepStatus.FAILURE in step_statuses:
-            return "🔴"  # Any failure = red
-        elif StepStatus.WARNING in step_statuses:
-            return "🟠"  # Warning = orange
-        elif StepStatus.UNKNOWN in step_statuses:
-            return "🟠"  # Unknown = orange
-        elif StepStatus.ONGOING in step_statuses:
-            return "🟢"  # Ongoing --> success
-        else:
-            return "🟢"  # All successful = green
-
-    except Exception as e:
-        logger.exception(f"Failed to check step statuses: {e}")
-        return "🔴"  # Error checking = red
-
-
-def _build_enhanced_notification(
-    artifact_dir: Path,
-    project: str,
-    finish_reason: FinishReason,
-    duration_str: str,
-    status: dict[str, Any],
-) -> str:
-    """Build enhanced notification with fournos job config and artifact links."""
-    fjob_project, fjob_args_str = _get_project_and_args(project)
-
-    # Check for job shutdown first (takes highest priority)
-    shutdown_status = _check_job_shutdown_status()
-    if shutdown_status and shutdown_status.get("is_aborted"):
-        status_emoji = "🛑"  # Abort status overrides everything
-    else:
-        # Check all step statuses for overall status emoji (takes priority over finish_reason)
-        status_emoji = _get_overall_status_from_steps(artifact_dir)
-
-    base_status = f"**{status_emoji} Execution of `{fjob_project}` {fjob_args_str} {status_emoji}**"
-
-    notification_parts = [base_status, "---"]
-
-    # Add job abort message right below overall status if applicable
-    if shutdown_status and shutdown_status.get("is_aborted"):
-        shutdown_value = shutdown_status.get("shutdown_value", "Stop")
-        notification_parts.append(f"🛑 **JOB ABORTED** - `spec.shutdown={shutdown_value}`")
-        notification_parts.append("")
-
-    execution_engine_config = _get_execution_engine_config()
-    if execution_engine_config:
-        notification_parts.append("**Execution Engine Configuration**")
-        notification_parts.append(execution_engine_config)
-
-    notification_success = True
-    try:
-        artifact_links, mlflow_run_url = _extract_artifact_links(status)
-        step_status = _process_step_status(artifact_dir, mlflow_run_url)
-        postprocess_status_links = _process_postprocess_status(mlflow_run_url)
-        test_status_section = _build_test_status_section(status, artifact_dir, mlflow_run_url)
-
-        if test_status_section:
-            notification_parts.append("")
-            notification_parts.append("---")
-            notification_parts.extend(test_status_section)
-
-        if artifact_links:
-            notification_parts.append("")
-            notification_parts.append("---")
-            notification_parts.append("**Artifact Links**")
-            notification_parts.extend([f"* {link}" for link in artifact_links])
-        else:
-            notification_parts.append("**Artifact Links:** No direct links available")
-
-        if step_status:
-            notification_parts.append("")
-            notification_parts.append("---")
-            notification_parts.append("**Step details**")
-            for link in step_status:
-                notification_parts.append(link)
-
-        if postprocess_status_links:
-            notification_parts.append("")
-            notification_parts.extend(postprocess_status_links)
-
-    except Exception as e:
-        logger.exception(f"Failed to build the extended notifications: {e}")
-        notification_parts.append("**Artifact Links:** Error extracting links")
-        notification_success = False
-
-    notification_parts.append("")
-    notification_parts.append("---")
-
-    return "\n".join(notification_parts), notification_success
-
-
-def _extract_test_labels_info(artifact_dir: Path, mlflow_run_url: str | None = None) -> list[str]:
-    """Extract test execution information from __test_labels__.yaml files.
-
-    Args:
-        artifact_dir: Directory to search for __test_labels__.yaml files
-        mlflow_run_url: Optional MLflow run URL for creating links
-
-    Returns:
-        List of formatted strings with test information (directory, labels, success, message)
-    """
-    from pathlib import Path
-
-    import yaml
-
-    test_info_lines = []
-
-    # Search for __test_labels__.yaml files recursively
-    test_labels_files = list(artifact_dir.glob("**/__test_labels__.yaml"))
-
-    if not test_labels_files:
-        return []
-
-    for test_labels_file in test_labels_files:
-        try:
-            with open(test_labels_file, encoding="utf-8") as f:
-                test_data = yaml.safe_load(f) or {}
-
-            # Extract directory relative to artifact_dir - use just the immediate directory name
-            relative_dir = test_labels_file.parent.relative_to(artifact_dir)
-            dir_name = relative_dir.name if relative_dir != Path(".") else "root"
-
-            # Extract completion info
-            completion = test_data.get("completion", {})
-            success = completion.get("success")
-            message = completion.get("message")
-
-            if message:
-                message = f": `{message}`"
-
-            # Format status
-            if success:
-                status_emoji = "✅"
-            elif success is False:
-                status_emoji = "❌"
-            else:
-                status_emoji = "❓"
-
-            # Create link to __test_labels__.yaml file if MLflow URL is available
-            if mlflow_run_url:
-                try:
-                    # Get the step directory name (relative to parent)
-                    step_dir_name = str(relative_dir)
-                    test_labels_url = _create_mlflow_file_url_for_step(
-                        mlflow_run_url, step_dir_name, "__test_labels__.yaml"
-                    )
-                    dir_link = f"[**{dir_name}**]({test_labels_url})"
-                except Exception as e:
-                    logger.warning(f"Failed to create MLflow link for {test_labels_file}: {e}")
-                    dir_link = f"**{dir_name}**"
-            else:
-                dir_link = f"**{dir_name}**"
-
-            test_info_lines.append(f"* {status_emoji} {dir_link}{message}")
-
-        except Exception as e:
-            test_info_lines.append(f"**{test_labels_file.name}**: Error reading file - {e}")
-
-    return test_info_lines
-
-
-def _extract_postprocess_status_info(artifact_dir: Path) -> list[str]:
-    """Extract post-processing status information from postprocess_status.yaml files.
-
-    Returns:
-        List of formatted strings with postprocess step status (success only, no details)
-    """
-    from pathlib import Path
-
-    import yaml
-
-    postprocess_info_lines = []
-
-    # Search for postprocess_status.yaml files recursively
-    postprocess_files = list(artifact_dir.glob("**/postprocess_status.yaml"))
-
-    if not postprocess_files:
-        return []
-
-    for postprocess_file in postprocess_files:
-        try:
-            with open(postprocess_file, encoding="utf-8") as f:
-                postprocess_data = yaml.safe_load(f) or {}
-
-            # Extract directory relative to artifact_dir
-            relative_dir = postprocess_file.parent.relative_to(artifact_dir)
-            dir_name = str(relative_dir) if relative_dir != Path(".") else "root"
-
-            # Extract overall status
-            overall_success = postprocess_data.get("success", False)
-            final_status = postprocess_data.get("final_status", "unknown")
-
-            # Extract individual step statuses
-            steps = postprocess_data.get("steps", [])
-            step_statuses = []
-
-            for step_dict in steps:
-                for step_name, step_data in step_dict.items():
-                    if isinstance(step_data, dict):
-                        status = step_data.get("status", "unknown")
-                        status_emoji = (
-                            "✅" if status == "success" else "❌" if status == "failed" else "⚪"
-                        )
-                        step_statuses.append(f"{step_name}:{status_emoji}")
-
-            # Format overall line
-            overall_emoji = "✅" if overall_success else "❌"
-            if step_statuses:
-                steps_str = " " + " ".join(step_statuses)
-            else:
-                steps_str = f" {final_status}"
-
-            postprocess_info_lines.append(f"**{dir_name}**: {overall_emoji}{steps_str}")
-
-        except Exception as e:
-            postprocess_info_lines.append(f"**{postprocess_file.name}**: Error reading file - {e}")
-
-    return postprocess_info_lines
-
-
-def _process_step_details(step_dir: Path, mlflow_run_url: str | None = None) -> dict[str, Any]:
-    """Process test labels and postprocess status for a single step directory."""
-    step_info = {"step_name": step_dir.name, "test_labels_info": [], "postprocess_info": []}
-
-    # Extract test labels for this specific step
-    try:
-        step_info["test_labels_info"] = _extract_test_labels_info(step_dir, mlflow_run_url)
-    except Exception as e:
-        logger.warning(f"Failed to extract test labels for step {step_dir.name}: {e}")
-
-    # Extract postprocess status for this specific step
-    try:
-        step_info["postprocess_info"] = _extract_postprocess_status_info(step_dir)
-    except Exception as e:
-        logger.warning(f"Failed to extract postprocess status for step {step_dir.name}: {e}")
-
-    return step_info
-
-
-def _build_test_status_section(
-    status: dict[str, Any], artifact_dir: Path, mlflow_run_url: str | None = None
-) -> list[str]:
-    """Build comprehensive test status section with step-by-step execution and post-processing overviews."""
-    try:
-        test_phase = status.get("test_phase", {})
-        if not test_phase:
-            return []
-
-        test_status = test_phase.get("phase", "UNKNOWN")
-        test_message = test_phase.get("message", "")
-
-        status_lines = [f"**Test status**: {test_status}"]
-
-        if test_message:
-            # Format message with blockquote-style prefix
-            status_lines.append(f"> {test_message}")
-
-        return status_lines
-
-    except Exception as e:
-        logger.warning(f"Failed to build test status section: {e}")
-        return []
-
-
-def _extract_finish_reason_from_status(status: dict[str, Any]) -> FinishReason:
-    """Extract finish reason from status object."""
-    # Check if any backend failed in the status
-    if not status:
-        return FinishReason.ERROR
-
-    # Look for backend results
-    backends = status.get("backends", {})
-    for backend_name, backend_result in backends.items():
-        # Check both explicit success flag and status field
-        if backend_result.get("success") is False or backend_result.get("status") not in (
-            None,
-            "success",
-        ):
-            logger.info(f"Backend {backend_name} failed, marking as error")
-            return FinishReason.ERROR
-
-    return FinishReason.SUCCESS
-
-
-def _extract_duration_from_status(status: dict[str, Any]) -> str:
-    """Extract duration from status object."""
-    # Look for duration in status
-    duration = status.get("duration")
-    if duration:
-        return f" after {duration}"
-    return ""
 
 
 def run_caliper_orchestration_export(*, artifact_dir: Path):
